requirements.txt
Flask~=3.0.2
flask_cors
python-dotenv~=1.0.1
google-generativeai
Werkzeug~=3.0.1
gunicorn
bcrypt~=4.1.2
langchain-community
boto3~=1.34.66
openai~=1.14.2
langchain~=0.1.12